+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Braintree Android SDK Release Notes
 
-## Unreleased
+## unreleased
 
 * UIComponents
     * Add Compose `CardFields` support to generate a premade credit card form for submitting credit card tokenize requests