+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Braintree Android SDK Release Notes
 
+## unreleased
+
+* UIComponents
+    * Add Compose `CardFields` support to generate a premade credit card form for submitting credit card tokenize requests
+
 ## 5.32.0 (2026-08-27)
 
 * GooglePay